refactor: 로그인, 회원가입 공통로직 분리
sign.js -> auth.js(공통)
